Split the archive reader and packer into two separate files
